refactor: extract main app shell
